+++ components/profile/EmploymentDialog.js
@@ -1,19 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import firebase from 'firebase/app';
-import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
-
-import CircularProgress from '@material-ui/core/CircularProgress';
-import Dialog from '@material-ui/core/Dialog';
-import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
-import SubmitSuccess from '../activityInput/SubmitSuccess';
-import { useAuth } from '../Auth';
-import { DialogContent, DialogTitle, DialogActions } from '../DialogComponents';
 
 const useStyles = makeStyles(theme => ({
   textField: {
@@ -54,67 +45,11 @@ const MONTHS = [
   'December',
 ];
 
-const initialState = {
-  title: undefined,
-  org: undefined,
-  startMonth: '',
-  endMonth: '',
-  startYear: undefined,
-  endYear: undefined,
-};
-
-function EmploymentDialog({ show, onClose, mode, items, itemIndex }) {
-  const formId = 'edit-profile';
+function EmploymentDialog({ handleChange, handleSubmit, values }) {
+  const formId = 'employmentItems';
   const classes = useStyles();
-  const { userDocRef } = useAuth();
-  const [success, setSuccess] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState();
-  const [values, setValues] = useState({});
 
-  const handleChange = event => {
-    setValues({
-      ...values,
-      [event.target.name]: event.target.value,
-    });
-  };
-
-  const reset = () => {
-    setValues({});
-    setError();
-    setLoading(false);
-    setSuccess(false);
-  };
-
-  useEffect(() => {
-    if (show && Object.keys(values) < 1) {
-      if (mode === UPDATE) {
-        const { start, end, ...rest } = items[itemIndex];
-        const [startMonth, startYear] = start.split(' ');
-        const [endMonth, endYear] = end.split(' ');
-        setValues(vals => ({
-          ...vals,
-          ...rest,
-          startYear,
-          startMonth,
-          endYear,
-          endMonth,
-        }));
-      } else {
-        setValues(vals => ({ ...vals, ...initialState }));
-      }
-    }
-    if (!show) {
-      reset();
-    }
-  }, [show]);
-
-  const handleClose = () => {
-    reset();
-    onClose();
-  };
-
-  const handleSubmit = async () => {
+  const onSubmit = () => {
     const { startMonth, startYear, endMonth, endYear, org, title } = values;
     const updatedItem = {
       start: `${startMonth} ${startYear}`,
@@ -123,207 +58,135 @@ function EmploymentDialog({ show, onClose, mode, items, itemIndex }) {
       title,
     };
 
-    const updatedItems =
-      mode === ADD
-        ? [...items, updatedItem]
-        : [...items.slice(0, itemIndex), updatedItem, ...items.slice(itemIndex + 1)];
-
-    setError();
-    setSuccess();
-    setLoading(true);
-    try {
-      await userDocRef.update({
-        lastUpdateTimestamp: firebase.firestore.FieldValue.serverTimestamp(),
-        'userProfile.employmentItems': updatedItems,
-      });
-      setSuccess(true);
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
+    handleSubmit(updatedItem);
   };
 
   return (
     <>
-      <Dialog
-        maxWidth="sm"
-        open={show}
-        aria-labelledby="edit-employment-dialog"
-        onExited={handleClose}
-      >
-        <DialogTitle id="edit-profile-dialog" onClose={onClose}>
-          <Typography variant="h6">Update Employment Experience</Typography>
-          <Typography variant="body1" color="textSecondary">
-            Lorem ipsum
-          </Typography>
-        </DialogTitle>
-        <DialogContent>
-          {loading && <CircularProgress />}
-          <SubmitSuccess message="Employment updated" show={success} />
-          {!(loading || success) && (
-            <form id={formId}>
-              <span>Job Title</span>
-              <TextField
-                className={classes.textField}
-                fullWidth
-                id={`${formId}-title`}
-                InputLabelProps={{ shrink: true }}
-                inputProps={{ name: 'title' }}
-                margin="normal"
-                onChange={handleChange}
-                placeholder="Enter the Job Title"
-                value={values.title}
-                variant="outlined"
-              />
-              <span>Company / Organization</span>
-              <TextField
-                className={classes.textField}
-                fullWidth
-                id={`${formId}-org`}
-                InputLabelProps={{ shrink: true }}
-                inputProps={{ name: 'org' }}
-                margin="normal"
-                onChange={handleChange}
-                placeholder="Enter the Company / Organization Name"
-                value={values.org}
-                variant="outlined"
-              />
-              <span>Dates</span>
-              <Grid container wrap="nowrap" spacing={1}>
-                <Grid item md={3}>
-                  <Select
-                    fullWidth
-                    margin="normal"
-                    variant="outlined"
-                    className={classes.textField}
-                    InputLabelProps={{ shrink: true }}
-                    inputProps={{ name: 'startMonth', placeholder: 'Month' }}
-                    id={`${formId}-start-month`}
-                    value={values.startMonth}
-                    onChange={handleChange}
-                    displayEmpty
-                  >
-                    <MenuItem value="" disabled>
-                      <span className={classes.placeholderText}>Month</span>
-                    </MenuItem>
-                    {MONTHS.map(month => (
-                      <MenuItem key={month} value={month}>
-                        {month}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </Grid>
-                <Grid item md={3}>
-                  <TextField
-                    fullWidth
-                    variant="outlined"
-                    className={classes.textField}
-                    id={`${formId}-start-year`}
-                    InputLabelProps={{ shrink: true }}
-                    inputProps={{ name: 'startYear' }}
-                    margin="normal"
-                    type="number"
-                    onChange={handleChange}
-                    placeholder="Year"
-                    value={values.startYear}
-                  />
-                </Grid>
-                <div className={classes.dash}>–</div>
-                <Grid item md={3}>
-                  <Select
-                    fullWidth
-                    margin="normal"
-                    variant="outlined"
-                    className={classes.textField}
-                    InputLabelProps={{ shrink: true }}
-                    inputProps={{ name: 'endMonth', placeholder: 'Month' }}
-                    id={`${formId}-end-month`}
-                    value={values.endMonth}
-                    onChange={handleChange}
-                    displayEmpty
-                  >
-                    <MenuItem value="" disabled>
-                      <span className={classes.placeholderText}>Month</span>
-                    </MenuItem>
-                    {MONTHS.map(month => (
-                      <MenuItem key={month} value={month}>
-                        {month}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </Grid>
-                <Grid item md={3}>
-                  <TextField
-                    fullWidth
-                    variant="outlined"
-                    className={classes.textField}
-                    id={`${formId}-end-year`}
-                    InputLabelProps={{ shrink: true }}
-                    inputProps={{ name: 'endYear' }}
-                    margin="normal"
-                    type="number"
-                    onChange={handleChange}
-                    placeholder="Year"
-                    value={values.endYear}
-                  />
-                </Grid>
-              </Grid>
-            </form>
-          )}
-          {error && (
-            <Typography color="error" variant="h4">
-              Error: {error}
-            </Typography>
-          )}
-        </DialogContent>
-        <DialogActions>
-          {!success && (
-            <Button
-              onClick={handleSubmit}
-              color="primary"
+      <form id={formId} onSubmit={onSubmit}>
+        <span>Job Title</span>
+        <TextField
+          className={classes.textField}
+          fullWidth
+          id={`${formId}-title`}
+          InputLabelProps={{ shrink: true }}
+          onChange={event => handleChange('title', event.target.value)}
+          placeholder="Enter the Job Title"
+          value={values.title}
+          variant="outlined"
+        />
+        <span>Company / Organization</span>
+        <TextField
+          className={classes.textField}
+          fullWidth
+          id={`${formId}-org`}
+          InputLabelProps={{ shrink: true }}
+          inputProps={{ name: 'org' }}
+          onChange={event => handleChange('org', event.target.value)}
+          placeholder="Enter the Company / Organization Name"
+          value={values.org}
+          variant="outlined"
+        />
+        <span>Dates</span>
+        <Grid container wrap="nowrap" spacing={1}>
+          <Grid item md={3}>
+            <Select
               fullWidth
-              size="large"
-              variant="contained"
+              variant="outlined"
+              className={classes.textField}
+              inputProps={{ name: 'startMonth', placeholder: 'Month' }}
+              id={`${formId}-start-month`}
+              value={values.startMonth || ''}
+              onChange={event => handleChange('startMonth', event.target.value)}
+              displayEmpty
             >
-              Save
-            </Button>
-          )}
-          {success && (
-            <Button
-              onClick={handleClose}
-              color="primary"
+              <MenuItem value="" disabled>
+                <span className={classes.placeholderText}>Month</span>
+              </MenuItem>
+              {MONTHS.map(startMonth => (
+                <MenuItem key={startMonth} value={startMonth}>
+                  {startMonth}
+                </MenuItem>
+              ))}
+            </Select>
+          </Grid>
+          <Grid item md={3}>
+            <TextField
               fullWidth
-              size="large"
-              variant="contained"
+              variant="outlined"
+              className={classes.textField}
+              id={`${formId}-start-year`}
+              inputProps={{ name: 'startYear' }}
+              type="number"
+              onChange={event => handleChange('startYear', event.target.value)}
+              placeholder="Year"
+              value={values.startYear}
+            />
+          </Grid>
+          <div className={classes.dash}>–</div>
+          <Grid item md={3}>
+            <Select
+              fullWidth
+              variant="outlined"
+              className={classes.textField}
+              inputProps={{ name: 'endMonth', placeholder: 'Month' }}
+              id={`${formId}-end-month`}
+              value={values.endMonth || ''}
+              onChange={event => handleChange('endMonth', event.target.value)}
+              displayEmpty
             >
-              Close
-            </Button>
-          )}
-        </DialogActions>
-      </Dialog>
+              <MenuItem value="" disabled>
+                <span className={classes.placeholderText}>Month</span>
+              </MenuItem>
+              {MONTHS.map(endMonth => (
+                <MenuItem key={endMonth} value={endMonth}>
+                  {endMonth}
+                </MenuItem>
+              ))}
+            </Select>
+          </Grid>
+          <Grid item md={3}>
+            <TextField
+              fullWidth
+              variant="outlined"
+              className={classes.textField}
+              id={`${formId}-end-year`}
+              InputLabelProps={{ shrink: true }}
+              inputProps={{ name: 'endYear' }}
+              type="number"
+              onChange={event => handleChange('endYear', event.target.value)}
+              placeholder="Year"
+              value={values.endYear}
+            />
+          </Grid>
+        </Grid>
+      </form>
     </>
   );
 }
 
 EmploymentDialog.propTypes = {
-  show: PropTypes.bool.isRequired,
-  onClose: PropTypes.func.isRequired,
-  mode: PropTypes.oneOf([ADD, UPDATE]).isRequired,
-  itemIndex: PropTypes.number,
-  items: PropTypes.arrayOf(
-    PropTypes.shape({
-      start: PropTypes.string,
-      end: PropTypes.string,
-      title: PropTypes.string,
-      org: PropTypes.string,
-    })
-  ),
+  handleChange: PropTypes.func.isRequired,
+  handleSubmit: PropTypes.func.isRequired,
+  values: PropTypes.shape({
+    startYear: PropTypes.string,
+    startMonth: PropTypes.string,
+    endYear: PropTypes.string,
+    endMonth: PropTypes.string,
+    title: PropTypes.string,
+    org: PropTypes.string,
+  }),
 };
 
 EmploymentDialog.defaultProps = {
-  itemIndex: null,
-  items: [],
+  values: {
+    title: undefined,
+    org: undefined,
+    startMonth: '',
+    endMonth: '',
+    startYear: undefined,
+    endYear: undefined,
+  },
 };
 
 export default EmploymentDialog;
